moved items around to handle new untitledQueryEditorInput location
