+++ src/sql/workbench/contrib/tasks/browser/tasksView.ts
@@ -27,7 +27,7 @@ import { TaskHistoryRenderer } from 'sql/workbench/contrib/tasks/browser/tasksRe
 import { TaskHistoryDataSource } from 'sql/workbench/contrib/tasks/browser/tasksDataSource';
 import { TaskHistoryController } from 'sql/workbench/contrib/tasks/browser/tasksController';
 import { TaskHistoryActionProvider } from 'sql/workbench/contrib/tasks/browser/tasksActionProvider';
-import { ITaskService } from 'sql/workbench/services/tasks/common/tasksService';
+import { ITaskService } from 'sql/workbench/services/tasks/browser/tasksService';
 import { TaskNode, TaskStatus } from 'sql/workbench/services/tasks/common/tasksNode';
 import { IErrorMessageService } from 'sql/platform/errorMessage/common/errorMessageService';
 import { IExpandableTree } from 'sql/workbench/services/objectExplorer/browser/treeUpdateUtils';
