+++ src/sql/workbench/test/electron-browser/api/mainThreadBackgroundTaskManagement.test.ts
@@ -6,7 +6,7 @@
 import * as azdata from 'azdata';
 import { Mock, It, Times } from 'typemoq';
 import { MainThreadBackgroundTaskManagement, TaskStatus } from 'sql/workbench/api/browser/mainThreadBackgroundTaskManagement';
-import { ITaskService } from 'sql/workbench/services/tasks/common/tasksService';
+import { ITaskService } from 'sql/workbench/services/tasks/browser/tasksService';
 import { IExtHostContext } from 'vs/workbench/api/common/extHost.protocol';
 import { TaskNode } from 'sql/workbench/services/tasks/common/tasksNode';
 import { Emitter } from 'vs/base/common/event';
